Refactored out into 2 seperate classes
